model-train-image/requirements.txt
joblib 
numpy 
pandas==1.4.4
scikit-learn
google-cloud-storage
evidently
xgboost
matplotlib
